pyproject.toml
[tool.poetry]
name = "redata-ext"
version = "0.0.1"
description = "`redata-ext` is a Meltano utility extension."
authors = [""]
license = "Apache 2.0"
packages = [
    { include = "redata_ext" }
    # uncomment this if you want to include static assets
    # { include = "files_redata-ext.*" }
]
# If you need to static assets with your extension, you can add them here.
# Under a directory path of "files_redata-ext/"
#include = [
#    "files_redata-ext/somefile.thing"
#]

[tool.poetry.dependencies]
python = "<4,>=3.8"
PyYAML = "^6.0.0"
click = "^8.1.3"
typer = "^0.6.1"
"meltano.edk"= {git = "https://github.com/meltano/edk.git", rev="main"}

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"
flake8 = "^3.9.0"
dbt = "^1.0.0.38.5"
re-data = "^0.11.0"

[[tool.poetry.source]]
name = "mirrors"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
priority = "primary"

[build-system]
requires = ["poetry-core>=1.0.8"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
redata_extension = 'redata_ext.main:app'
redata_invoker = 'redata_ext.pass_through:pass_through_cli'
